# PYTHON-1721 Improve GridFS file download performance