… doc/changelog.rst
… test/test_session.py