Jira: https://jira.mongodb.org/browse/PYTHON-1721

This change uses a single cursor to download all the chunks in a GridFS file instead of using individual find_one operations to read each chunk. I've also refactored the code a bit to add earlier detection of truncated chunks.  

The performance benefit of using a single cursor will vary but it should always be an improvement. 

Python 2.7 local server: negligible benefit (latency is ~0.09ms):
```
$ mongo-python-driver git:(PYTHON-1721-master-final) TEST_PATH=driver-performance-test-data python test/performance/perf_test.py TestGridFsDownload # Local repl 1 node, cursor
Running TestGridFsDownload. MEDIAN=0.322574853897
.{
    "results": [
        {
            "name": "TestGridFsDownload",
            "results": {
                "1": {
                    "ops_per_sec": 162532197.9273853
                }
            }
        }
    ]
}

----------------------------------------------------------------------
Ran 1 test in 35.708s

OK
$ mongo-python-driver git:(PYTHON-1721-master-final) gcm
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
$ mongo-python-driver git:(master) TEST_PATH=driver-performance-test-data python test/performance/perf_test.py TestGridFsDownload # Local repl 1 node, find_one
Running TestGridFsDownload. MEDIAN=0.326012134552
.{
    "results": [
        {
            "name": "TestGridFsDownload",
            "results": {
                "1": {
                    "ops_per_sec": 160818553.80029458
                }
            }
        }
    ]
}

----------------------------------------------------------------------
Ran 1 test in 33.828s

OK
```


Python 3.7 local server: 15% improvement:
```
$ mongo-python-driver git:(PYTHON-1721-master-final) TEST_PATH=driver-performance-test-data python3.7 test/performance/perf_test.py TestGridFsDownload # Local repl 1 node, cursor
Running TestGridFsDownload. MEDIAN=0.25183119000000076
.{
    "results": [
        {
            "name": "TestGridFsDownload",
            "results": {
                "1": {
                    "ops_per_sec": 208190256.33798516
                }
            }
        }
    ]
}

----------------------------------------------------------------------
Ran 1 test in 29.609s

OK
$ mongo-python-driver git:(PYTHON-1721-master-final) gcm
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
$ mongo-python-driver git:(master) TEST_PATH=driver-performance-test-data python3.7 test/performance/perf_test.py TestGridFsDownload # Local repl 1 node, cursor
Running TestGridFsDownload. MEDIAN=0.29158426900000034
.{
    "results": [
        {
            "name": "TestGridFsDownload",
            "results": {
                "1": {
                    "ops_per_sec": 179806682.23222953
                }
            }
        }
    ]
}

----------------------------------------------------------------------
Ran 1 test in 30.502s

OK
```

Now, once we run the benchmark against a server with some real world latency, using a cursor becomes much more advantageous. Let's test with an Atlas (M10) cluster.

Python 2.7, Atlas 4.0.6 replica set, 33% speed up (latency is ~5ms):
```
$ mongo-python-driver git:(PYTHON-1721-master-final) <ATLAS_ENV_VARS> TEST_PATH=driver-performance-test-data python test/performance/perf_test.py TestGridFsDownload # Atlas find cursor
test/performance/perf_test.py:114: UserWarning: Test timed out, completed 83 iterations.
  warnings.warn('Test timed out, completed %s iterations.' % i)
Running TestGridFsDownload. MEDIAN=3.35720396042
.{
    "results": [
        {
            "name": "TestGridFsDownload",
            "results": {
                "1": {
                    "ops_per_sec": 15616805.120610315
                }
            }
        }
    ]
}

----------------------------------------------------------------------
Ran 1 test in 313.563s

OK
$ mongo-python-driver git:(PYTHON-1721-master-final) gcm
M	test/__init__.py
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
$ mongo-python-driver git:(master) <ATLAS_ENV_VARS> TEST_PATH=driver-performance-test-data python test/performance/perf_test.py TestGridFsDownload # Atlas find_one
test/performance/perf_test.py:114: UserWarning: Test timed out, completed 65 iterations.
  warnings.warn('Test timed out, completed %s iterations.' % i)
Running TestGridFsDownload. MEDIAN=4.43294596672
.{
    "results": [
        {
            "name": "TestGridFsDownload",
            "results": {
                "1": {
                    "ops_per_sec": 11827078.514739025
                }
            }
        }
    ]
}

----------------------------------------------------------------------
Ran 1 test in 311.489s

OK
```

Python 3.7, Atlas 4.0.6 replica set: Again ~33% speed up:
```
$ mongo-python-driver git:(PYTHON-1721-master-final) <ATLAS_ENV_VARS> TEST_PATH=driver-performance-test-data python3.7 test/performance/perf_test.py TestGridFsDownload # Atlas find cursor
test/performance/perf_test.py:114: UserWarning: Test timed out, completed 91 iterations.
  warnings.warn('Test timed out, completed %s iterations.' % i)
Running TestGridFsDownload. MEDIAN=3.1946222420000083
.{
    "results": [
        {
            "name": "TestGridFsDownload",
            "results": {
                "1": {
                    "ops_per_sec": 16411580.471303768
                }
            }
        }
    ]
}

----------------------------------------------------------------------
Ran 1 test in 317.615s

OK
$ mongo-python-driver git:(PYTHON-1721-master-final) gcm
M	test/__init__.py
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
$ mongo-python-driver git:(master) <ATLAS_ENV_VARS> TEST_PATH=driver-performance-test-data python3.7 test/performance/perf_test.py TestGridFsDownload # Atlas find_one
test/performance/perf_test.py:114: UserWarning: Test timed out, completed 65 iterations.
  warnings.warn('Test timed out, completed %s iterations.' % i)
Running TestGridFsDownload. MEDIAN=4.296203145000021
.{
    "results": [
        {
            "name": "TestGridFsDownload",
            "results": {
                "1": {
                    "ops_per_sec": 12203519.766288832
                }
            }
        }
    ]
}

----------------------------------------------------------------------
Ran 1 test in 313.310s

OK
```