@ShaneHarvey 2019-03-13
Merged to master: https://github.com/mongodb/mongo-python-driver/commit/956fd92e826cd73eeaf383007654cfd02d9c6cb3 and v3.8: https://github.com/mongodb/mongo-python-driver/commit/8069e132329fed15fc56372631dce7bdabcdfc32.